# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8050",
    "basePath": "/",
    "paths": {
        "/v1/cat": {
            "get": {
                "description": "finds all the cats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Finds all the cats",
                "operationId": "find-all-cats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Cat"
                            }
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "creates a cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a cat",
                "operationId": "create-cat",
                "parameters": [
                    {
                        "description": "Cat parameters",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/CreateCatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Cat"
                        }
                    }
                }
            }
        },
        "/v1/cat/{id}": {
            "get": {
                "description": "get cat by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Finds a cat by its id",
                "operationId": "get-cat-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Cat"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateCatRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.Cat": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string",
                    "example": "Black"
                },
                "id": {
                    "type": "string",
                    "example": "8d6dde41-d23a-49a5-a9d7-6cbba52ec7fc"
                },
                "name": {
                    "type": "string",
                    "example": "Ryuk"
                }
            }
        }
    }
}